+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.2
 
 * Fix incorrect escaping of notebook names ([#566](https://github.com/databrickslabs/terraform-provider-databricks/pull/566))
+* Added support for spot instances on Azure ([#571](https://github.com/databrickslabs/terraform-provider-databricks/pull/571))
 
 ## 0.3.1